Begun World, Cell and Agent Scripts
Have begun World script (World being a 2D array of Cell objects) which includes a method to create the Cell objects for the array and populate Sugar/Spice mountains (this method I have only just started though, very early stages).
Agent script does not yet include all instance variables and only method so far is one which allows the agent to 'know' the world it is living in.